// src/app/api/models/service-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ServiceResponseDto {
  basePrice: number;
  category?: string;
  createdAt: string;
  description?: string;
  duration?: number;
  id: number;
  name: string;
  notes?: string;
  ownerId: number;
  status: 'ACTIVE' | 'INACTIVE';
  thumbnailUrl?: string;
  updatedAt: string;
}