// src/app/api/models/update-service-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateServiceDto {
  basePrice?: number;
  category?: string;
  description?: string;

  /**
   * Duration in minutes
   */
  duration?: number;
  name?: string;
  notes?: string;
  status?: 'ACTIVE' | 'INACTIVE';
}